GETP-123 fix: Redux Slice 와 ThunkAction 사이 순환참조 버그 해결
