+++ src/store/slice/signup.slice.ts
@@ -1,7 +1,3 @@
-import { authService } from "@/services/auth/auth.service";
-import { MemberType } from "@/services/auth/auth.types";
-
-import { RootDispatch, store } from "../store";
 import { PayloadAction, createSlice } from "@reduxjs/toolkit";
 
 export interface ISignUpState {
@@ -65,33 +61,3 @@ const signUpSlice = createSlice({
 
 export const signUpAction = signUpSlice.actions;
 export const signUpReducer = signUpSlice.reducer;
-
-export const signUpThunkAction = () => {
-    return async (dispatch: RootDispatch, getState: typeof store.getState) => {
-        // 회원가입 POST 요청
-
-        const {
-            signUpMemberType,
-            signUpEmailValue,
-            signUpPasswordValue,
-            signUpAgreementValue,
-            signUpVerificationCode,
-        } = getState().signUp;
-
-        await authService.signUp({
-            memberType: signUpMemberType as MemberType,
-            email: signUpEmailValue as string,
-            password: signUpPasswordValue as string,
-            verificationCode: signUpVerificationCode as string,
-            serviceTerms: [
-                {
-                    tag: "GET-P 서비스 약관",
-                    agreed: signUpAgreementValue,
-                },
-            ],
-        });
-
-        dispatch(signUpAction.nextSignUpSection());
-        dispatch(signUpAction.initalizeState());
-    };
-};
