+++ src/store/thunk/auth.thunk.ts
@@ -1,0 +1,5 @@
+import { RootDispatch } from "../store";
+
+export const reissueTokenThunkAction = () => {
+    return async (dispatch: RootDispatch) => {};
+};
